1.计时器代码优化

+++ app/src/main/java/com/maunc/toolbox/chronograph/data/ChronographData.kt
@@ -2,6 +2,6 @@ package com.maunc.toolbox.chronograph.data
 
 data class ChronographData(
     var index: Int,//名次
-    var time: Float,//使用时间
-    var gapTime: Float,//与上一名时间差距
+    var time: Long,//使用时间
+    var gapTime: Long,//与上一名时间差距
 )
